requirements.txt
numpy==2.2.3
pandas==2.2.3
requests==2.32.3
apache-airflow==2.10.5
notebook==7.3.2
mlflow==2.20.4
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.1
plotly-express==0.4.1
docker==7.1.0
dvc==3.59.1
dvc[s3]
dagshub==0.5.9
python-dotenv
